backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
motor==3.1.2
pymongo==4.3.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
fastapi-mail==1.1.1
jinja2==3.1.2
python-dotenv==1.0.0
pandas==2.0.2
numpy==1.24.3
scikit-learn==1.2.2
plotly==5.14.1
joblib==1.2.0
pytest==7.3.1
httpx==0.24.0
geopy==2.3.0
requests==2.31.0
aiohttp==3.8.5 